requirements.txt
fastapi
uvicorn
pypdf
requests
google-generativeai
sentence-transformers
chromadb
